Renamed all .h files to .hpp
