+++ src/bgnbd_expectation.hpp
