+++ src/clv_vectorized.hpp
