+++ src/ggomnbd_expectation.cpp
@@ -1,6 +1,6 @@
 #include <RcppArmadillo.h>
 #include <math.h>
-#include "ggomnbd_LL.h"
+#include "ggomnbd_LL.hpp"
 
 
 // integrand <- function(tau)  {tau * exp(b*tau)  *((beta + exp(b*tau) - 1)^(-(s+1)))}
