+++ src/pnbd_DERT.cpp
@@ -1,7 +1,7 @@
 #include <RcppArmadillo.h>
 #include <math.h>
-#include "pnbd_LL_ind.h"
-#include "clv_vectorized.h"
+#include "pnbd_LL_ind.hpp"
+#include "clv_vectorized.hpp"
 
 //' @name pnbd_DERT
 //'
