+++ src/pnbd_PAlive.cpp
@@ -2,8 +2,8 @@
 #include <math.h>
 #include <vector>
 
-#include "clv_vectorized.h"
-#include "pnbd_LL_ind.h"
+#include "clv_vectorized.hpp"
+#include "pnbd_LL_ind.hpp"
 
 //' @name pnbd_PAlive
 //'
