+++ src/pnbd_expectation.cpp
@@ -1,7 +1,7 @@
 #include <RcppArmadillo.h>
 #include <math.h>
-#include "pnbd_expectation.h"
-#include "clv_vectorized.h"
+#include "pnbd_expectation.hpp"
+#include "clv_vectorized.hpp"
 
 //' @name pnbd_expectation
 //' @title Pareto/NBD: Unconditional Expectation
